refactor(infra): move warp decoder logic
